003_jobs/requirements.txt
# Google Cloud dependencies
google-cloud-logging==3.8.0
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0

# Core dependencies
requests==2.31.0
pandas==2.1.4
numpy==1.25.2

# Utility libraries
python-dateutil==2.8.2
pytz==2023.3

# Development dependencies (optional for Cloud Run)
pytest==7.4.3
black==23.12.0
flake8==6.1.0